refactor: 패키지분류
