+++ tomcat/src/main/java/org/apache/coyote/handler/LoginRequestHandler.java
@@ -18,8 +18,8 @@ import org.apache.coyote.HttpResponse;
 import org.apache.coyote.RequestHandler;
 import org.apache.coyote.http11.Http11Method;
 import org.apache.coyote.http11.Http11Processor;
-import org.apache.coyote.http11.Http11Response;
 import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.response.Http11Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
