+++ tomcat/src/main/java/org/apache/coyote/handler/NotFoundHandler.java
@@ -7,8 +7,8 @@ import java.io.IOException;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
 import org.apache.coyote.RequestHandler;
-import org.apache.coyote.http11.Http11Response;
 import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.response.Http11Response;
 
 public class NotFoundHandler implements RequestHandler {
