+++ tomcat/src/main/java/org/apache/coyote/handler/SignupRequestHandler.java
@@ -16,8 +16,8 @@ import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
 import org.apache.coyote.RequestHandler;
 import org.apache.coyote.http11.Http11Method;
-import org.apache.coyote.http11.Http11Response;
 import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.response.Http11Response;
 
 public class SignupRequestHandler implements RequestHandler {
