+++ tomcat/src/main/java/org/apache/coyote/http11/request/Http11RequestHeaders.java
@@ -1,9 +1,10 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.coyote.http11.HttpCookie;
 
 public class Http11RequestHeaders {
